requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
supabase==1.0.3
python-jose==3.3.0
requests==2.28.2
Werkzeug==2.2.3
Flask-Login==0.6.2 